[FA-1] Refactor directorys
